Move into config_validator_util directory
