+++ google/cloud/forseti/scanner/scanners/config_validator_util/data_models/data_model_requirements_map.py
